+++ app/jobs/disco_app/synchronise_resources_job.rb
@@ -1,0 +1,3 @@
+class DiscoApp::SynchroniseResourcesJob < DiscoApp::ShopJob
+  include DiscoApp::Concerns::SynchroniseResourcesJob
+end